third_party/WebKit/LayoutTests/external/wpt/referrer-policy/same-origin/attr-referrer/same-origin/http-https/script-tag/same-origin-insecure.swap-origin-redirect.http-expected.txt

This is a testharness.js-based test.
FAIL The referrer URL is omitted when a
                                 document served over http requires an https
                                 sub-resource via script-tag using the attr-referrer
                                 delivery method with swap-origin-redirect and when
                                 the target request is same-origin. assert_equals: Reported Referrer URL is 'omitted'. expected (undefined) undefined but got (string) "http://web-platform.test:8001/referrer-policy/same-origin/attr-referrer/same-origin/http-https/script-tag/same-origin-insecure.swap-origin-redirect.http.html"
Harness: the test ran to completion.
